move files back to reduce diff
